requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymysql==1.1.1
redis==6.4.0
pytest==8.4.2
httpx==0.28.1
bcrypt==4.2.0
PyJWT==2.9.0
